Change csc to u8 in enr and metadata

+++ beacon_node/lighthouse_network/src/peer_manager/peerdb/peer_info.rs
@@ -375,7 +375,10 @@ impl<E: EthSpec> PeerInfo<E> {
             if let Ok(custody_subnet_count) = meta_data.custody_subnet_count() {
                 let custody_subnets = DataColumnSubnetId::compute_custody_subnets::<E>(
                     node_id.raw().into(),
-                    std::cmp::min(*custody_subnet_count, spec.data_column_sidecar_subnet_count),
+                    std::cmp::min(
+                        *custody_subnet_count as u64,
+                        spec.data_column_sidecar_subnet_count,
+                    ),
                     spec,
                 )
                 .collect::<HashSet<_>>();
